fix: decouple Chat Tools with config libs
